+++ docs/md/releasenotes.md
@@ -7,6 +7,7 @@
 - Fix potential `StringIndexOutOfBoundsException` in `KotlinCallableReferenceInitializer`. (`T5927`)
 - Fix potential `NullPointerException` in `KotlinInterClassPropertyReferenceInitializer`. (`T6138`)
 - Fix wrong offset for complementary branch instruction when widening branch instructions in `InstructionWriter`. (`T5721`)
+- Fix potential `ClassFormatError` due to adding multiple annotation attributes when processing Kotlin code.
 
 ## Version 8.0.1